+++ clang/docs/ReleaseNotes.rst
@@ -200,7 +200,9 @@ X86 Support in Clang
   implies -mtune=<cpu>. -mtune=generic is the default with no -march or -mtune
   specified.
 
-- Support for feature ``HRESET`` has been added.
+- Support for ``HRESET`` instructions has been added.
+
+- Support for ``UINTR`` instructions has been added.
 
 Internal API Changes
 --------------------